+++ src/SumatraPDF.cpp
@@ -8329,6 +8329,10 @@ static LRESULT FrameOnCommand(MainWindow* win, HWND hwnd, UINT msg, WPARAM wp, L
             FindToggleMatchCase(win);
             break;
 
+        case CmdFindToggleMatchWholeWord:
+            FindToggleMatchWholeWord(win);
+            break;
+
         case CmdFindNextSel:
             FindSelection(win, TextSearch::Direction::Forward);
             break;